Added a templates folder, a handler and a routes file outside boilerplate's folder, with the '/secure/' url as an example of how to add your own application files.
